+++ packages/pangea-sdk/PangeaCyber.Net/Audit/Models/DownloadFormat.cs
@@ -1,0 +1,19 @@
+using Newtonsoft.Json;
+using System.Runtime.Serialization;
+
+namespace PangeaCyber.Net.Audit.Models
+{
+    ///
+    [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
+    public enum DownloadFormat
+    {
+        ///
+        [EnumMember(Value = "json")]
+        JSON,
+
+        ///
+        [EnumMember(Value = "csv")]
+        CSV,
+
+    }
+}