+++ packages/pangea-sdk/CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Documentation for client builders and constructors.
+- Audit /download_results endpoint support
+
+### Fixed
+
+- Put to presigned url. It should just put file in raw, not in form format.
 
 ### Changed